Separate general MQTT from Linux
