+++ microflo/linux_mqtt_main.hpp
@@ -6,7 +6,7 @@
 #define DEBUG 1
 
 #include "microflo.h"
-#include "mqtt.hpp"
+#include "linux_mqtt.hpp"
 
 #include "main_maps.h" // generated graph metadata
 #include "microflo.hpp"
